requirements.txt
djangorestframework==3.14.0
Django==5.0
psycopg2-binary==2.9.9
python-decouple==3.8
pillow==10.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
weasyprint==61.2
pydyf==0.10.0